#11 - implemented asto blob
