+++ src/test/java/com/artipie/docker/storage/AstoRepoITCase.java
@@ -27,7 +27,7 @@ package com.artipie.docker.storage;
 import com.artipie.asto.FileStorage;
 import com.artipie.docker.Repo;
 import com.artipie.docker.RepoName;
-import com.artipie.docker.manifest.ManifestRef;
+import com.artipie.docker.ref.ManifestRef;
 import java.nio.file.Path;
 import javax.json.JsonObject;
 import org.hamcrest.MatcherAssert;
